задание 5
