+++ .eslintrc.yml
@@ -1,6 +1,9 @@
+parserOptions:
+  sourceType: "module"
+
 env:
   es6: true
   browser: true
   commonjs: true
 
-extends: 'htmlacademy/es6'
+extends: "htmlacademy/es6"
