+++ public/index.html
@@ -45,6 +45,6 @@
         <p>130 291 movies inside</p>
       </section>
     </footer>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
