+++ src/components/film-extra-list.js
@@ -1,0 +1,9 @@
+const createFilmExtraList = text => {
+  return `<section class="films-list--extra">
+      <h2 class="films-list__title">${text}</h2>
+
+      <div class="films-list__container"></div>
+    </section>`;
+};
+
+export { createFilmExtraList };
