+++ src/components/filter.js
@@ -1,0 +1,9 @@
+const createFilterTemplate = () => {
+  return `<ul class="sort">
+    <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
+    <li><a href="#" class="sort__button">Sort by date</a></li>
+    <li><a href="#" class="sort__button">Sort by rating</a></li>
+  </ul>`;
+};
+
+export { createFilterTemplate };
