+++ src/components/show-more.js
@@ -1,0 +1,5 @@
+const createShowMoreTemplate = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
+
+export { createShowMoreTemplate };
